+++ src/lightning/pytorch/plugins/precision/hpu.py
@@ -15,9 +15,9 @@ from typing import cast, Literal, Optional, Union
 
 from typing_extensions import get_args
 
+from lightning.pytorch.accelerators.hpu import _HPU_AVAILABLE
 from lightning.pytorch.plugins.precision.precision_plugin import PrecisionPlugin
 from lightning.pytorch.utilities.exceptions import MisconfigurationException
-from lightning.pytorch.utilities.imports import _HPU_AVAILABLE
 
 if _HPU_AVAILABLE:
     from habana_frameworks.torch.hpex import hmp
