+++ src/lightning/pytorch/utilities/__init__.py
@@ -17,10 +17,10 @@ import numpy
 
 from lightning.fabric.utilities import LightningEnum  # noqa: F401
 from lightning.fabric.utilities import move_data_to_device  # noqa: F401
+from lightning.pytorch.accelerators.hpu import _HPU_AVAILABLE  # noqa: F401
 from lightning.pytorch.utilities.enums import GradClipAlgorithmType  # noqa: F401
 from lightning.pytorch.utilities.grads import grad_norm  # noqa: F401
 from lightning.pytorch.utilities.imports import (  # noqa: F401
-    _HPU_AVAILABLE,
     _OMEGACONF_AVAILABLE,
     _TORCH_QUANTIZE_AVAILABLE,
     _TORCHVISION_AVAILABLE,
